+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### UNRELEASED
 
+### v2.0.4 (2020-08-03)
+
 - MOB-835: Fix refresh token issue
 
 ### v2.0.3 (2017-05-30)